+++ doc/changelog.rst
@@ -10,6 +10,7 @@ New features, enhancements:
 
 * Shortcuts can be user defined.
 * Events now have a progress attribute.
+* Find feature for categories with Ctrl+F when mouse in category tree.
 
 Bug fixes: